server/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
cached-property==1.5.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
referencing==0.34.0
rpds-py==0.18.0
SQLAlchemy==2.0.29
sqlalchemy-orm==1.2.10
typing-inspect==0.9.0
typing_extensions==4.10.0
Werkzeug==3.0.1
